requirements.txt
torch
torchvision
transformers
Pillow
nltk
pandas
numpy
matplotlib
tqdm
pycocotools